(setup-redesign): Redesign InputNodeName (#710)
