+++ src/i18n/langs/de.json
@@ -216,6 +216,13 @@
       "shutdown_info_disconnect_power": "Anschließend 5 Sekunden warten und dann vom Strom trennen."
     },
     "setup": {
+      "input_node": {
+        "err_only_chars": "Der Name der Node darf nur aus Buchstaben und Zahlen bestehen.",
+        "err_max_length": "Der Name der Node darf maximal 32 Zeichen lang sein.",
+        "err_min_length": "Der Name der Node muss mindestens 4 Zeichen lang sein.",
+        "err_empty": "Der Name der Node darf nicht leer sein.",
+        "nodename_name": "Name der Node"
+      },
       "blockchain_delete": "Blockchain löschen",
       "blockchain_found_long": "Auf deiner HDD/SSD wurde eine existierende Blockchain gefunden.",
       "blockchain_found_short": "Blockchain auf HDD/SSD gefunden",
@@ -247,12 +254,6 @@
       "migrate_umbrel": "Willst du von UMBREL auf RaspiBlitz migrieren?",
       "migrateblitz": "Migriere zu RaspiBlitz",
       "no_and_shutdown": "Nein und herunterfahren",
-      "nodename_err_chars": "Der Name der Node darf nur aus Buchstaben und Zahlen bestehen.",
-      "nodename_err_max_length": "Der Name der Node darf maximal 32 Zeichen lang sein.",
-      "nodename_err_min_length": "Der Name der Node muss mindestens 4 Zeichen lang sein.",
-      "nodename_error_empty": "Der Name der Node darf nicht leer sein.",
-      "nodename_input": "Gib einen Namen für die Node ein",
-      "nodename_name": "Name der Node",
       "ok": "OK",
       "other_options": "Andere Optionen",
       "password_a_details": "Nutze dieses Passwort für den Login zum RaspiBlitz.",
